backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
fastf1==3.3.5
pandas==2.1.3
python-dotenv==1.0.0
redis==5.0.1
supabase==2.0.3
requests==2.31.0
python-multipart==0.0.6
httpx==0.25.2
numpy==1.26.2
python-json-logger==2.0.7 